Convert dials_data to pytest plugin
* define pytest hooks in pytest11.py
* disable and deprecate previous hooks
